A ting bug?

+++ Helicopter.cpp
@@ -11,9 +11,9 @@ void Helicopter::fly(int headwind, int minutes){
     if (extraWeight < 0) 
         extraWeight = 0;
     if (headwind >= 40){
-        delta_fuel = (float) minutes *(0.4+0.01*(float) extraWeight);
+        delta_fuel = (float) minutes * (0.4+0.01*(float) extraWeight);
     } else {
-        delta_fuel = minutes*(0.4+0.01* (float) extraWeight);
+        delta_fuel = (float) minutes * (0.4+0.01*(float) extraWeight);
     }
     if (Helicopter::fuel-delta_fuel < 20.0){
         return;
